@@ -20,19 +20,19 @@
     "test": "mocha"
   },
   "devDependencies": {
-    "@types/node": "^16.11.55",
+    "@types/node": "^16.18.3",
     "@types/xml2js": "^0.4.11",
-    "@typescript-eslint/eslint-plugin": "5.33.0",
-    "@typescript-eslint/parser": "5.35.1",
-    "mocha": "^7.1.1",
-    "node-gyp": ">=1.0.0",
+    "@typescript-eslint/eslint-plugin": "5.42.0",
+    "@typescript-eslint/parser": "5.42.0",
+    "mocha": "10.1.0",
+    "node-gyp": ">=9.3.0",
     "nyc": "^15.0.1",
-    "typescript": "4.8.2",
-    "eslint": "8.11.0",
+    "typescript": "4.8.4",
+    "eslint": "8.26.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react-app": "7.0.0",
-    "eslint-plugin-jest": "^26.1.2",
-    "eslint-plugin-prettier": "4.0.0",
+    "eslint-config-react-app": "7.0.1",
+    "eslint-plugin-jest": "^27.1.4",
+    "eslint-plugin-prettier": "4.2.1",
     "prettier": "^2.4.1"
   },
   "dependencies": {
